[chore] basic website with some documents on getting started
