+++ .gitignore
@@ -134,3 +134,5 @@ dist
 # TernJS port file
 
 .tern-port
+
+.DS_Store
