+++ package.json
@@ -20,11 +20,17 @@
     "url": "https://github.com/gabrielcsapo/miette/issues"
   },
   "homepage": "https://github.com/gabrielcsapo/miette#readme",
+  "dependencies": {
+    "error-stack-parser": "^2.0.6",
+    "jest": "^27.1.1",
+    "tslint": "^6.1.3",
+    "typescript": "^4.4.2"
+  },
   "devDependencies": {
     "@types/jest": "^27.0.1",
-    "@types/node": "^16.7.13",
-    "@typescript-eslint/eslint-plugin": "^4.31.1",
-    "@typescript-eslint/parser": "^4.31.1",
+    "@types/node": "^16.10.1",
+    "@typescript-eslint/eslint-plugin": "^4.31.2",
+    "@typescript-eslint/parser": "^4.31.2",
     "chalk": "^4.1.2",
     "eslint": "^7.32.0",
     "eslint-config-prettier": "^8.3.0",
@@ -34,12 +40,6 @@
     "prettier": "^2.4.1",
     "ts-jest": "^27.0.5",
     "ts-node": "^10.2.1"
-  },
-  "dependencies": {
-    "error-stack-parser": "^2.0.6",
-    "jest": "^27.1.1",
-    "tslint": "^6.1.3",
-    "typescript": "^4.4.2"
   },
   "workspaces": [
     "website"
