--- website/docs/tutorial-basics/_category_.json
@@ -1,4 +1,0 @@
-{
-  "label": "Tutorial - Basics",
-  "position": 2
-}
