--- website/docs/tutorial-extras/_category_.json
@@ -1,4 +1,0 @@
-{
-  "label": "Tutorial - Extras",
-  "position": 3
-}
