+++ website/src/pages/index.module.css
@@ -3,6 +3,19 @@
  * and scoped locally.
  */
 
+ .features {
+  display: flex;
+  align-items: center;
+  padding: 2rem 0;
+  width: 100%;
+}
+
+.featureSvg {
+  height: 200px;
+  width: 200px;
+}
+
+
 .heroBanner {
   padding: 4rem 0;
   text-align: center;
